requirements.txt
Flask==3.0.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
Flask-Mail==0.9.1
Flask-Bcrypt==1.0.1
